Only download web2py if it doesn't already exist

+++ scripts/install.sh
@@ -20,24 +20,27 @@ cp ${NGINX_CONFIG_PATH}/default-web2py.conf /config/etc/nginx-web2py/conf.d/defa
 apt-get install -y --no-install-recommends git-core
 pip3 install gitpython
 
-## Install Web2py
-mkdir -p /opt/tmp
-cd /opt/tmp
-git clone https://github.com/web2py/web2py.git --depth 1 --branch v2.21.1 --single-branch web2py
-cd web2py
-git submodule update --init --recursive
-cd ../../
-
-if [ "${WEB2PY_MIN}" == true ]; then
-  cd tmp/web2py
-  python3 scripts/make_min_web2py.py ../../tmp/web2py-min
-  mv ../../tmp/web2py-min ../../web2py
+## Install Web2py into /opt/web2py if it is not already been installed there by previous layers
+if [ ! -d /opt/web2py ]; then
+  mkdir -p /opt/tmp
+  cd /opt/tmp
+  git clone https://github.com/web2py/web2py.git --depth 1 --branch v2.21.1 --single-branch web2py
+  cd web2py
+  git submodule update --init --recursive
   cd ../../
-else
-  mv tmp/web2py web2py
+
+  if [ "${WEB2PY_MIN}" == true ]; then
+    cd tmp/web2py
+    python3 scripts/make_min_web2py.py ../../tmp/web2py-min
+    mv ../../tmp/web2py-min ../../web2py
+    cd ../../
+  else
+    mv tmp/web2py web2py
+  fi
+
+  rm -rf tmp
 fi
 
-rm -rf tmp
 mv web2py/handlers/wsgihandler.py web2py/wsgihandler.py
 
 cp /build/bin/web2py-setpass /opt/web2py
